Renamed package index to graph
